// src/Type/product.type.ts
import { IMarket } from "./market.type";


export interface IProduct {
  id: number;
  slug: string;
  nameBn: string;
  category: string;
  categoryNameBn: string;
  categoryIcon: string;
  unit: string;
  image: string;
  today: number;
  yesterday: number;
  lastWeek: number;
  lastMonth: number;

  change: {
    dir: "up" | "down";
    pct: number;
  };

  markets: IMarket[];
}